rename BingoCard to BingoTile and extract it to new file
