+++ chringo/lib/main.dart
@@ -2,6 +2,8 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 
+import 'bingo_tile.dart';
+
 void main() => runApp(MyApp());
 
 final dummySnapshot = [
@@ -16,7 +18,7 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      title: 'Baby Names',
+      title: 'CHRINGO',
       home: MyHomePage(),
     );
   }
@@ -62,41 +64,6 @@ class _MyHomePageState extends State<MyHomePage> {
         return new BingoTile(words[index]);
       }),
     );
-  }
-}
-
-class BingoTile extends StatefulWidget {
-  final String word;
-  BingoTile(this.word);
-
-  @override
-  _BingoTileState createState() => _BingoTileState(word);
-}
-
-class _BingoTileState extends State {
-  bool _isChecked = false;
-  final String word;
-  _BingoTileState(this.word);
-
-  @override
-  Widget build(BuildContext context) {
-    return GestureDetector(
-      child: Card(
-        color: _isChecked ? Colors.green : Colors.white,
-        child: Center(
-          child: Text(
-            word,
-            style: Theme.of(context).textTheme.subtitle,
-            textAlign: TextAlign.center,
-          ),
-        ),
-      ),
-      onTap: _toggleChecked,
-    );
-  }
-
-  void _toggleChecked() {
-    setState(() => _isChecked = !_isChecked);
   }
 }
